389s: Fix random ssl breakage
Depending on, I  don't know, the state of the moon or something, pk12util
will sometimes ignore the -n argument and name the cert after the CN.
Depending on the state of the sun or something, 389-ds will deem to
consider this being a valid certificate or not (the docs say that it only
consideres certificates named after nsSSLPersonalitySSL, which would have
been the CN, but the docs are definitely wrong). Fix this by always renaming
the certificate to 'Server-Cert', forcing nsSSLPersonalitySSL to the same value
and ignoring errors.
It should be against some kind of law to use this certificate database format...